Fix bug that can cause SCEWL controller to hang
Fix the following bug:
When no SCEWL messages are being sent to a given SED, that SED's SCEWL
controller may enter the read_msg function because its antenna received
a message that was addressed to some other SED. At that point, read_msg
will not return until that SED receives a message addressed to it or a
broadcast. In many cases, this could result in the SCEWL controller
staying in read_msg for a long time, preventing the main loop from
running during that time. It may appear to the user that the SCEWL
controller has hung/frozen.