Uninitialized variable (res) in scewl_recv
In the scewl_recv function, if the uninitialized variable "res" happens to pick up the value "2" from previously used RAM, the error catch for SCEWL_NO_MSG on line 184 generates a false NO MSG and returns an error code.  The calling program receives the error which prevents the data transfer when the data is actually valid.  The problem persists until a different RAM location is used when calling scewl_recv.  Initializing the res variable to SCEWL_OK enum fixes the issue.